Replace the m_archetype_table hashmap type to std::unordered_map
The m_archetype_table contains strong references to entities in the ECS framework.
Entities can be constructed and destroyed frequently during the runtime; a memory-pooled node-chaining-based hash map would be a better fit, since it performs better than open-addressing-based maps if the number of elements is not small and if the insertion and deletion are frequent.
In a typical ECS framework, an entity is a handle or just an integer representing an entity associated with an archetype. However, smart pointers are considered entities in the Frogman Engine ECS framework; the smart pointers are integral type handles with automatic memory management behavior since pointers are technically integers.